add Shopify provider
