+++ src/app/_components/buy-keycard-dialog.tsx
@@ -12,7 +12,6 @@ import { useController, useForm } from 'react-hook-form'
 import type { SubmitHandler } from 'react-hook-form'
 import { Shopify, shopifySchema } from '../_api/validation/shopify'
 import { KEYCARD_BUNDLES } from '../_constants/shopify/products'
-import { useShopifyUTMParams } from '../_hooks/use-shopify-utm-params'
 import {
   Check,
   Close,
@@ -24,6 +23,7 @@ import {
   Recommended,
   World,
 } from '../_icons'
+import { useShopifyUTMParamsContext } from '../_providers/shopify-utm-params-provider'
 import { createCart } from '../actions'
 import { Button } from './button'
 import { Form } from './form/form'
@@ -73,7 +73,7 @@ export const BuyKeycardDialog = (props: Props) => {
   const [open, setOpen] = useState(false)
 
   const router = useRouter()
-  const utmParams = useShopifyUTMParams()
+  const utmParams = useShopifyUTMParamsContext()
 
   const onSubmit: SubmitHandler<Shopify> = async data => {
     const quantity = data.quantity
