+++ src/app/_providers/index.tsx
@@ -1,0 +1,11 @@
+'use client'
+
+import { ShopifyUTMParamsProvider } from './shopify-utm-params-provider'
+
+type Props = {
+  children: React.ReactNode
+}
+
+export const Providers = ({ children }: Props) => {
+  return <ShopifyUTMParamsProvider>{children}</ShopifyUTMParamsProvider>
+}
