+++ src/app/layout.tsx
@@ -5,6 +5,7 @@ import './globals.css'
 import { Footer } from '~components/footer'
 import { NavBar } from '~components/nav-bar'
 import { NavBarMobile } from '~components/nav-bar-mobile'
+import { Providers } from './_providers'
 
 const lora = Lora({
   variable: '--font-lora',
@@ -40,12 +41,14 @@ export default function RootLayout({ children }: Props) {
           'bg-dark-100 p-2 font-inter text-white-100 antialiased',
         )}
       >
-        <NavBar />
-        <NavBarMobile />
-        <div className="flex w-full justify-center overflow-x-hidden">
-          <div className="w-full max-w-[1512px]">{children}</div>
-        </div>
-        <Footer />
+        <Providers>
+          <NavBar />
+          <NavBarMobile />
+          <div className="flex w-full justify-center overflow-x-hidden">
+            <div className="w-full max-w-[1512px]">{children}</div>
+          </div>
+          <Footer />
+        </Providers>
       </body>
     </html>
   )
